+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.14.2
+
+- @iov/socket: Add 10 second timeout to `SocketWrapper`. Value can be configured
+  using an optional constructor argument.
+- @iov/socket: `StreamingSocket` now times out after 10 seconds.
+
 ## 0.14.1
 
 - @iov/core: Add `JsonRpcSigningServer`.
@@ -24,6 +30,12 @@
   `parseJsonRpcResponse` is removed.
 - @iov/keycontrol: Export `ReadonlyWallet`.
 - @iov/keycontrol: Deprecate `HdPaths.simpleAddress`.
+
+## 0.13.7
+
+- @iov/socket: Add 10 second timeout to `SocketWrapper`. Value can be configured
+  using an optional constructor argument.
+- @iov/socket: `StreamingSocket` now times out after 10 seconds.
 
 ## 0.13.6